+++ common/src/main/java/net/mehvahdjukaar/every_compat/api/AbstractSimpleEntrySet.java
@@ -3,6 +3,7 @@ package net.mehvahdjukaar.every_compat.api;
 import com.google.common.base.Suppliers;
 import com.google.common.collect.ArrayListMultimap;
 import com.google.common.collect.Multimap;
+import com.google.gson.JsonObject;
 import com.mojang.datafixers.util.Pair;
 import net.mehvahdjukaar.every_compat.EveryCompat;
 import net.mehvahdjukaar.every_compat.configs.ModEntriesConfigs;
@@ -15,6 +16,7 @@ import net.mehvahdjukaar.moonlight.api.platform.PlatHelper;
 import net.mehvahdjukaar.moonlight.api.platform.RegHelper;
 import net.mehvahdjukaar.moonlight.api.resources.BlockTypeResTransformer;
 import net.mehvahdjukaar.moonlight.api.resources.RPUtils;
+import net.mehvahdjukaar.moonlight.api.resources.ResType;
 import net.mehvahdjukaar.moonlight.api.resources.SimpleTagBuilder;
 import net.mehvahdjukaar.moonlight.api.resources.pack.DynClientResourcesGenerator;
 import net.mehvahdjukaar.moonlight.api.resources.pack.DynamicDataPack;
@@ -31,6 +33,7 @@ import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.resources.ResourceLocation;
+import net.minecraft.server.packs.resources.Resource;
 import net.minecraft.server.packs.resources.ResourceManager;
 import net.minecraft.tags.TagKey;
 import net.minecraft.world.item.BlockItem;
@@ -39,6 +42,7 @@ import net.minecraft.world.item.Item;
 import net.minecraft.world.level.block.Block;
 import org.jetbrains.annotations.Nullable;
 
+import java.io.InputStream;
 import java.util.*;
 import java.util.function.BiFunction;
 import java.util.function.Consumer;
@@ -75,7 +79,6 @@ public abstract class AbstractSimpleEntrySet<T extends BlockType, B extends Bloc
     protected final Map<ResourceLocation, Set<ResourceKey<?>>> tags = new HashMap<>();
     protected final Set<Supplier<ResourceLocation>> recipeLocations = new HashSet<>();
     protected final Set<TextureInfo> textures = new HashSet<>();
-    @Nullable
     protected final BiFunction<T, ResourceManager, Pair<List<Palette>, @Nullable AnimationMetadataSection>> paletteSupplier;
     @Nullable
     protected final Consumer<BlockTypeResTransformer<T>> extraModelTransform;
@@ -89,7 +92,7 @@ public abstract class AbstractSimpleEntrySet<T extends BlockType, B extends Bloc
                                      Supplier<T> baseType,
                                      Supplier<ResourceKey<CreativeModeTab>> tab,
                                      TabAddMode tabMode,
-                                     @Nullable BiFunction<T, ResourceManager, Pair<List<Palette>, @Nullable AnimationMetadataSection>> paletteSupplier,
+                                     BiFunction<T, ResourceManager, Pair<List<Palette>, @Nullable AnimationMetadataSection>> paletteSupplier,
                                      @Nullable Consumer<BlockTypeResTransformer<T>> extraTransform,
                                      boolean mergePalette, boolean copyTint,
                                      Predicate<T> condition) {
@@ -272,12 +275,14 @@ public abstract class AbstractSimpleEntrySet<T extends BlockType, B extends Bloc
 
             Multimap<ResourceLocation, TextureInfo> infoPerTextures = ArrayListMultimap.create();
 
+            /// Adding multiple textures from one block into Respriter without/with mask & infoPerTextures
             for (var textureInfo : textures) {
                 ResourceLocation textureId = textureInfo.texture();
 
                 try {
                     ResourceLocation maskId = textureInfo.mask();
                     TextureImage main = TextureImage.open(manager, textureId);
+                    main.getMetadata();
 
                     infoPerTextures.put(textureId, textureInfo);
 
@@ -325,72 +330,79 @@ public abstract class AbstractSimpleEntrySet<T extends BlockType, B extends Bloc
             for (var e : partialRespriters.entrySet()) {
                 respriters.put(e.getKey(), Respriter.ofPalette(e.getValue(), globalPalette));
             }
-
+            /// Swapping out the old palettes of the texture with new plattes
             for (var entry : getDefaultEntries().entrySet()) {
                 var b = entry.getValue();
                 T w = entry.getKey();
                 // skips disabled ones
                 // actually we dont otherwise we get missign texture log spam. TODO: replace models with empty dummy instead
                 // if (!ModConfigs.isEntryEnabled(w, b)) continue;
-
                 ResourceLocation blockId = Utils.getID(b);
 
-                List<Palette> targetPalette = null;
-                AnimationMetadataSection animation = null;
-                if (paletteSupplier != null) {
-                    var pal = paletteSupplier.apply(w, manager);
-                    animation = pal.getSecond();
-                    targetPalette = pal.getFirst();
-                } else {
-                    var m = w.mainChild();
-                    Block mainWoodTypeBlock = null;
-                    if (m instanceof Block bb) mainWoodTypeBlock = bb;
-                    else if (m instanceof BlockItem bii) mainWoodTypeBlock = bii.getBlock();
-                    if (mainWoodTypeBlock == null) {
-                        throw new UnsupportedOperationException("You need to provide a palette supplier for non wood type based blocks");
-                    }
+                var pal = paletteSupplier.apply(w, manager); // return the texture of: WoodType: Planks, StoneType: stone, LeavesType: leaves
+                AnimationMetadataSection targetAnimation = pal.getSecond();
+                List<Palette> targetPalette = pal.getFirst();
 
-                    try (TextureImage plankTexture = TextureImage.open(manager,
-                            RPUtils.findFirstBlockTextureLocation(manager, mainWoodTypeBlock))) {
-                        targetPalette = Palette.fromAnimatedImage(plankTexture);
-                        animation = plankTexture.getMetadata();
-                    } catch (Exception ignored) {
-                    }
-                }
                 if (targetPalette == null) {
                     EveryCompat.LOGGER.error("Could not get texture palette for block {} : ", b);
                     continue;
                 }
-                AnimationMetadataSection finalAnimation = animation;
-                List<Palette> finalTargetPalette = targetPalette;
 
                 //sanity check to verity that palette isn't changed. can be removed
-                int oldSize = finalTargetPalette.get(0).size();
+                int oldSize = targetPalette.get(0).size();
 
+                /// Creating new Path to add the new textures via the resources
                 for (var re : respriters.entrySet()) {
-                    if (oldSize != finalTargetPalette.get(0).size()) {
+                    if (oldSize != targetPalette.get(0).size()) {
                         throw new RuntimeException("This should not happen");
                     }
                     ResourceLocation oldTextureId = re.getKey();
                     String oldPath = oldTextureId.getPath();
 
-                    String newPath = BlockTypeResTransformer.replaceTypeNoNamespace(oldPath, w,
-                            blockId, baseType.get().getTypeName());
-                    String newId = new ResourceLocation(blockId.getNamespace(), newPath).toString();
+                    // boatload's texture path has 2 folder
+                    String newPath = (oldPath.startsWith("entity/") && module.modId.equals("boatload"))
+                            ? BlockTypeResTransformer.replaceFullGenericType(oldPath, w, blockId, baseType.get().getTypeName(), null, 2)
+                            // Default
+                            : BlockTypeResTransformer.replaceTypeNoNamespace(oldPath, w, blockId, baseType.get().getTypeName());
+
+                    String newId = "";
 
                     boolean isOnAtlas = true;
 
+                    /// Adding the textures to the resource
                     for (var info : infoPerTextures.get(oldTextureId)) {
                         if (info != null) {
-                            if (info.keepNamespace()) {
-                                newId = oldTextureId.withPath(newPath).toString();
+                            if (info.keepNamespace()) newId = oldTextureId.withPath(newPath).toString();
+                            else newId = new ResourceLocation(blockId.getNamespace(), newPath).toString();
+
+                            if (newId.isEmpty()) {
+                                EveryCompat.LOGGER.error("The path of new texture is empty for: {}", info.texture());
+                                continue;
                             }
+
                             isOnAtlas = info.onAtlas();
+
+                            /// TEMP: do not remove this until the mcmeta problem is fixed.
+                            if (info.copyMCMETA()) {
+                                ResourceLocation mcmetaLoc = ResType.MCMETA.getPath(oldTextureId);
+                                Optional<Resource> getMCMETA = manager.getResource(mcmetaLoc);
+
+                                if (getMCMETA.isPresent()) {
+                                    InputStream mcmetaStream = getMCMETA.get().open();
+                                    JsonObject mcmetaFile = RPUtils.deserializeJson(mcmetaStream);
+
+                                    // Adding to the resources next to newtextures
+                                    handler.dynamicPack.addJson(ResourceLocation.tryParse(newId), mcmetaFile, ResType.MCMETA);
+                                    mcmetaStream.close();
+                                }
+                                else
+                                    handler.getLogger().error("The MCMETA file may no longer existing, check @ {}", mcmetaLoc);
+                            }
                         }
 
                         Respriter respriter = re.getValue();
 
-                        Supplier<TextureImage> textureSupplier = () -> respriter.recolorWithAnimation(finalTargetPalette, finalAnimation);
+                        Supplier<TextureImage> textureSupplier = () -> respriter.recolorWithAnimation(targetPalette, targetAnimation);
                         textureSupplier = postProcessTexture(w, newId, manager, textureSupplier);
 
                         handler.addTextureIfNotPresent(manager, newId, textureSupplier, isOnAtlas);
@@ -420,6 +432,25 @@ public abstract class AbstractSimpleEntrySet<T extends BlockType, B extends Bloc
         return textureSupplier;
     }
 
+    private static Pair<List<Palette>, @Nullable AnimationMetadataSection> getPaletteFromMainChild(BlockType w, ResourceManager manager) {
+        var mainChild = w.mainChild();
+        Block mainWoodTypeBlock = null;
+        if (mainChild instanceof Block bb) mainWoodTypeBlock = bb;
+        else if (mainChild instanceof BlockItem bii) mainWoodTypeBlock = bii.getBlock();
+        if (mainWoodTypeBlock == null) {
+            throw new UnsupportedOperationException("You need to provide a palette supplier for non block main child");
+        }
+
+        try (TextureImage plankTexture = TextureImage.open(manager,
+                RPUtils.findFirstBlockTextureLocation(manager, mainWoodTypeBlock))) {
+            var targetPalette = Palette.fromAnimatedImage(plankTexture);
+            var animation = plankTexture.getMetadata();
+            return Pair.of(targetPalette, animation);
+        } catch (Exception ignored) {
+        }
+        return Pair.of(null, null);
+    }
+
 
     @SuppressWarnings("unchecked")
     protected static class Builder<BL extends Builder<BL, T, B, I>, T extends BlockType, B extends Block, I extends Item> {
@@ -430,8 +461,7 @@ public abstract class AbstractSimpleEntrySet<T extends BlockType, B extends Bloc
         protected final String prefix;
         protected Supplier<ResourceKey<CreativeModeTab>> tab = null;
         protected TabAddMode tabMode = TabAddMode.AFTER_SAME_TYPE;
-        @Nullable
-        protected BiFunction<T, ResourceManager, Pair<List<Palette>, @Nullable AnimationMetadataSection>> palette = null;
+        protected BiFunction<T, ResourceManager, Pair<List<Palette>, @Nullable AnimationMetadataSection>> palette = AbstractSimpleEntrySet::getPaletteFromMainChild;
         protected final Map<ResourceLocation, Set<ResourceKey<?>>> tags = new HashMap<>();
         protected final Set<Supplier<ResourceLocation>> recipes = new HashSet<>();
         protected final Set<TextureInfo> textures = new HashSet<>();
@@ -466,8 +496,13 @@ public abstract class AbstractSimpleEntrySet<T extends BlockType, B extends Bloc
         }
 
         //exclusive with addCondition
-        public BL requiresFromMap(Map<T, ?> entrySet) {
-            this.addCondition(blockType -> Objects.nonNull(entrySet.get(blockType)));
+        public BL requiresFromMap(Map<?, ?>... entrySets) {
+            this.addCondition(blockType -> {
+                for (Map<?, ?> entrySet : entrySets) {
+                    if (Objects.isNull(entrySet.get(blockType))) return false;
+                }
+                return true;
+            });
             return (BL) this;
         }
 
@@ -479,7 +514,7 @@ public abstract class AbstractSimpleEntrySet<T extends BlockType, B extends Bloc
             regexBuilder.append(modId).append(":(");
             for (int i = 0; i < ids.length; i++) {
                 regexBuilder.append(ids[i]);
-                if ( i != (ids.length - 1) ) regexBuilder.append("|"); // Don't append "|" to the last word's
+                if (i != (ids.length - 1)) regexBuilder.append("|"); // Don't append "|" to the last word's
             }
             regexBuilder.append(")");
 
@@ -590,7 +625,8 @@ public abstract class AbstractSimpleEntrySet<T extends BlockType, B extends Bloc
         }
 
         public BL createPaletteFromOak() {
-            return createPaletteFromOak(p -> {});
+            return createPaletteFromOak(p -> {
+            });
         }
 
         public BL createPaletteFromChild(Consumer<Palette> paletteTransform, String childKey) {
@@ -598,11 +634,13 @@ public abstract class AbstractSimpleEntrySet<T extends BlockType, B extends Bloc
         }
 
         public BL createPaletteFromChild(String childKey, Predicate<String> whichSide) {
-            return createPaletteFromChild(p -> {}, childKey, whichSide);
+            return createPaletteFromChild(p -> {
+            }, childKey, whichSide);
         }
 
         public BL createPaletteFromChild(String childKey) {
-            return createPaletteFromChild(p -> {}, childKey, null);
+            return createPaletteFromChild(p -> {
+            }, childKey, null);
         }
 
         public BL createPaletteFromChild(Consumer<Palette> paletteTransform, String childKey, Predicate<String> whichSide) {