+++ common/src/main/java/net/mehvahdjukaar/every_compat/api/SimpleEntrySet.java
@@ -72,7 +72,7 @@ public class SimpleEntrySet<T extends BlockType, B extends Block> extends Abstra
                           @Nullable TriFunction<T, B, Item.Properties, Item> itemFactory,
                           @Nullable SimpleEntrySet.ITileHolder<?> tileFactory,
                           @Nullable Object renderType,
-                          @Nullable BiFunction<T, ResourceManager, Pair<List<Palette>, @Nullable AnimationMetadataSection>> paletteSupplier,
+                          BiFunction<T, ResourceManager, Pair<List<Palette>, @Nullable AnimationMetadataSection>> paletteSupplier,
                           @Nullable Consumer<BlockTypeResTransformer<T>> extraTransform,
                           boolean mergedPalette, boolean copyTint,
                           Predicate<T> condition) {
@@ -277,11 +277,12 @@ public class SimpleEntrySet<T extends BlockType, B extends Block> extends Abstra
     protected BlockTypeResTransformer<T> makeLootTableTransformer(SimpleModule module, ResourceManager manager) {
         String oldTypeName = baseType.get().getTypeName();
         return BlockTypeResTransformer.<T>create(module.modId, manager)
+                // Modifying the JSON filenames & path
+                .setIDModifier((text, blockId, type) ->
+                        BlockTypeResTransformer.replaceFullGenericType(text, type, blockId, oldTypeName, null, 2))
+                // Modifying the JSON files' content
                 .addModifier((text, blockId, type) ->
-                        BlockTypeResTransformer.replaceFullGenericType(text, type, blockId, oldTypeName, module.modId, 2))
-                .addModifier((text, blockId, type) ->
-                        ResourcesUtils.convertItemIDinText(text, baseType.get(), type))
-                .IDReplaceType(oldTypeName);
+                        ResourcesUtils.convertItemIDinText(text, baseType.get(), type));
     }
 
     //ok...